Numpy error found!
It looks like the problem isn't numpy, but the calculation method.
If the differences between `modelFile` and `benchFile` are negitive,
`numpy.amax()` will find zero to be the maximum (as it should). To
calculate the max errors, you need to take the absolute value of the
errors. If the +/- is important, then more care will have to be taken
with these calculations.

Also, summing the errors without taking the absolute value can lead to
errors if there are positive and negitive errors -- errors can cancel
and the total error will be under reported.

It's probably worth reviewing what information, exactly, we want from
these error variables.