LIVV now runs on titan
Added a setup_*.bash script for titan and hopper to load the modules.
This will represent the last known working modules on the system.

Also, set the matplotlib backend to `Agg` so that it doesn't try and use
`GTKAgg` which isn't available on these systems (shouldn't matter for
our plots since we don't `show()` them)

This is in reference to #2