Added new scheduler scaffolding and bundle directory.
This commit provides a decent view of what is to come.  With the bundle directory we will be able to provide standardized input file sets for common run configurations (equivalent to the standard options that we have previously supplied via command line).  This will allow us to easily extend the code to different ISM/dycore combinations as well as reduce the ammount of hard coded strings left in the code.  This also removes the need for test objects and we can simply think of a test as an operation on data.  Ideally this will make it more intuitive for others to use and extend our code.  It also provides the benefit of making data verification a simpler process.  It is now evident where to look for each piece of analysis is stored within LIVV.