Don't use the Python Pandas module on Yellowstone
This takes the pandas module out of the yellowstone setup script, which
allows the default modules to be loaded (instead of versioned ones) and
reduces the number of needed modules. Pandas will then be installed
automatically via setuptools, which will take a little to build the first
time it's run for each user.