Streamlined access to bundle-specific code.
Workflow will be moved to reduce options and make the loading of functionality based on model-dycore combo built into the input structure.  To use LIVV in an executable fashion now the command

-------------------------------------------------------------------
                      __   _____   ___   ____    _ __
                     / /  /  _/ | / / | / / /__ (_) /_
                    / /___/ / | |/ /| |/ /  '_// / __/
                   /____/___/ |___/ |___/_/\_\/_/\__/

                   Land Ice Verification & Validation
-------------------------------------------------------------------
  Load modules: python, ncl, nco, python_matplotlib, hdf5, netcdf,
                python_numpy, and python_netcdf4 for best results.

  Current run: 02-14-2016 22:37:42
  User: bzq
  OS Type: Linux 4.4.1-2-ARCH
  Machine: katha

Beginning Dependency Checks........
    Checking for external libraries....
      Found numpy!
      Found netCDF4!
      Found pandas!
      Found matplotlib!
Okay!
Setting up environment....
-------------------------------------------------------------------
Previous output data found in output directory!
   Backing up data to:
   /home/bzq/livv-dev/LIVV/vv_02-14-2016_02142016_221418
-------------------------------------------------------------------

where  and  are paths to the output directories of a bundle.  The leaf of the bundle paths must specify one of the included bundle-compatibility packages given in the  directory of LIVV.  This commit specifically sets up the infrastructure for passing off bundle-specific functionality.  For a concrete example, this commit makes the implementation for verification of CISM-Glissade under the  bundle;  to use this pass  two paths that point to a  directory that have the new output directory structure of CISM's BATS capability.

Future work includes moving the  function out of  and into  and separating the functionality that is CISM-Glissade specific out of .  Currently this progress does not extend to comparing different bundle implementations in the verification aspect, but can easily be extended by using  and  in the  method.

Additionally there has been some further reductions in the old code by removing validation infrastructure no longer in use.