Some style updates for the performance plots.
The scaling plots have been changed such that there will be a partially
transparent fill between the min and max lines, the legend will be
placed in the best location, and the colors have been updated.

The timing breakdown plots now have a bar for both bench and test, the
axis labels hav ebeen updated, the legend only have a single set of
labels and the legend has been moved outside of all the bar plots.