Immediately choose a matplotlib backend
On HPC machines, there will be no $DISPLAY, and so matplotlib needs to
use the 'agg' backend. This needs to be set before pyplot is imported
for the first time, or anything uses matplotlib. Therefore, let's just
do it as the very first thing LIVVkit does.

Also, this solves a Gtk display error on my linux machine.