Updating TexHelper.
The TexHelper module now contains an object that can be used to build Latex output at runtime.  A new option '--make-tex' has been added to allow for its use.  When invoked the standard output is still produced, but inside of the output directory there will be an additional 'tex' directory that contains the generated files.  A master file called  has each of the individual components '\input'd.  It can be built with the standard 'pdflatex' command.