Use wvel when wvel_ho doesn't exist
`wvel_ho` was an outdated and redundant diagnostic variable for CISM
which has been removed. ISMIP-HOM tests have historically used it for
comparisons, but they have been updated to use `wvel`.

LIVVkit now will try to find `wvel_ho` in the ISMIP-HOM f datasets for
CISM-glissade and if it cannot find it, will use `wvel`.

This is to maintain backwards compatibility for datasets produced with
`wvel_ho`, but can likely be dropped in the future.