Update CSS
Some minor changes to the CSS which brings back the glacier image and
leaves the font size as the system default.

The initial font size was previously enlarged because of the small
default font size on some screens with extremely high resolution.
However, this makes the initial font size too large on many standard and
low res displays. Using the system default seem to be the better option
overall.