Always use the same color for timers in timing breakdown plots
In the timing breakdown plots of the performance component, if a timer,
or timers, appeared the 'test' but not 'bench' dataset, then the same
timers (across the bar charts) could be assigned different colors.

This commit enforces using the same colors across by chart by assigning
colors based on the union of both timer sets.