Update the validation extension import process
LIVVkit will now be able to import any extension python module, no
matter where it is. Previously, the modules needed to be specified as an
import statement, and the module needed to be found on the python path.
For example, the validation template module is specified like:

    livvkit.components.validation_tests.template.template

Now, the module can be specified as an import statement or a valid path
to a python module file (specified either relative to your current
working directory or absolutely). For example, these three forms work:

1. livvkit.components.validation_tests.template.template
2. livvkit/components/validation_tests/template/template.py
3. /home/USER/code/livvkit/components/validation_tests/template/template.py

Additionally, a bug in the options parser was fixed, and the lines
around error messages are now made of exclamation points to stand out
more.