Update the way LIVVkit handles validation extension summaries
LIVVkit will now accept and display a summary table for each extension
so that they can report independent summary data (numerics,
verification, and performance all had standardized summary information,
which likely won't be true for the extensions).

The javascript function to draw the navigation menu has been updated to
handle this, and a new `ValSummary` element was added to livvkit so that
the summary title could have a user specified name.

The built-in validation tests have been updated to work with this new
method.