Update docs theme and directory structure
This adds in the previously ignored auto-generated sources directory.
These files are auto-generated by the generate_docs.sh script, but only
need to be generated if the LIVVkit source changes. These should be
version controlled so the entire documentation is available and can be
generated with just a `make html` command.