Changes to make documentation build with no warnings/errors
Changed the `CISM-glissade` bundle to `CISM_glissade` (and all
references throughout code/docs) because package names with a `-` are
not valid in python, and sphinx actually imports each package to
generate the docs.

Removed the unused `CISM-albany` bundle and the `js/__init__.py` file in
the resources component because they were viewed as empty packages by by
Sphinx and aren't needed for LIVVkit's functionality.

Removed the copy command in `generate_docs.sh` and directly include the
generated javascript documentation in the API doc. Also, removed the
unused jsdoc `files.rst` and `index.rst` files.

Included a `README.jst` in the docs directory to describe how to build
the documentation and included it in the development doc.

Also, added a `distclean` option to the Makefile which will clean the built
website and remove the auto-generated files (helpful for testing
documentation).