Add class for temporarily changing sys.path to import extension sub modules
As discussed in #5 , LIVVkit's method to import extension modules
doesn't allow for extension "packages" which have multiple *.py files.
For example, an extension may look like:

EXT_DIR
   -- ext.py
   -- SUB_PACKAGE
        -- __init__.py
        -- f1.py
        -- f2.py

but I can't get importlib set up in such a way as to allow the
SUB_PACKAGE to be imported elegantly from within ext.py (let alone in a
python 2 and 3 compatible way).

So, one workaround is to add EXT_DIR to the system path, but this will
possibly clutter the path and lead to name clashes if many extension are
run at the same time.

This adds a class for temporarily adding EXT_DIR to the path, and
removing once the needed imports are complete, which will keep the
sys.path cleaner and less likely for classes.

However, this isn't an ideal solution and eventually it'd be better to
figure out how to do this via importlib.