Better extension module import failure messages
This de-convolutes extension import failures so that internal module
import failures (e.g., missing internal deps) don't look like a missing
extension module.

Also, if the user is using conda, and there is a ext YAML file
describing the additional dependencies, it will suggest a conda command
to update the current environment with those dependencies.

fixes #6 and is related to #7