Remove license header in HTML files
This fixes the HTML warning, and subsequent reloading of the page, for
having the charset declared too deep in the HTML files.

Because the HTML files are intended to be viewed as rendered, and will
likely only be viewed in code by devs that will have seen this message
multiple times, there is little need to have a license header in HTML
files.