Remove numpy import in setup.py
Because `pip` is now the reccomended way to install packages in *all*
python versions, using `python setup.py install` is no longer expected
or support. Therefore, we no longer need to try and import numpy, and pip
install it upon failure, to prevent a direct build from souce that is
prone to failure.