move livv.py into livvkit.__main__.py
In an effort to make LIVVkit have better machine interoperability,
LIVVkit changed the script setup definition to an entry-point setup.
This, however, meant that the livv.py script was not being installed
properly because it wasn't a part of the livvkit package.

By moving the livv CLI into LIVVkit and setuptools, the `livv` script will
be created at package install time. For developers, this means that the
`livv` or `livv.py` script no longer exists in a git checkout, so either
installing a development (and editable) version of liv via
`pip install -e .` (recommended) or using livvkit like `python -m
livvkit [OPTIONS]` will now be required.