Redirect forked processes stdout and stderr to log files
This prevents cluttering of the terminal with out or order process
errors and preserves the output for later.

This method was chosen to extensions can still use `print` to provide
terminal output without having to worry about logs.

Once the processes are all finished, all the output is extracted from
the log files and is displaced in the terminal