Define a base element class for new-style elements
This begins the conversion to each each LIVVkit element providing a
_repr_STYLE function, and the LIVVkit output will be built based on that
style.

Current styles include JSON, HTML, LaTeX -- LIVVkit will output an HTML
website and/or a LaTeX document (uncompiled). Along side the HTML and/or
LaTeX, a JSON data representation will be created.

Ideally, with just the JSON, LIVVkit should be able to create an HTML
website, LATEX document, and/or any future provided styles.