Convert vhtable element to class-style element
The vhtable (table with vertical and horizontal headers) element has
been subsumed into the Table class with the addition of an index option.
if index=True, the header column will either use the Pandas DataFrame
index or number the table rows. If index=False (default) you'll get a
normal table without a header column. If index=Iterable then the header
column will use the iter-values to label the table rows.