Convert bit_for_bit element to class-style element
The bit-for-bit table (table with bit-for-bit difference plots in final
column) element has been converted to a BitForBit class. LIVVkit
produces the correct b4b tables on the output website, but does not yet
take full advantage of the class methods.

Also adds a B4BImage and a NAImage which take place of the bit-for-bit
difference plots (of nothing-plot) or stand is as a placeholder if the
bit-for-bit comparison can't be run (e.g., variable missing from either
test or reference data), respectively.

Further:
* Does some cleanup of the CSS while ading styling for the images
  in the bit-for-bit table.
* Adds the LIVVkit favicon.ico to all the templates (when'd that
  dissapear?).
* Removes unused common.js functions
* Fixes Gallery Type
* Fixes the URL generation/path specification for Image elements