Update LIVVkit tab elements to new Class style elements
The tab element dictionary has been converted to a Tabs class.  LIVVkit
produces the correct tabs on the output website, but does not yet
take full advantage of the class methods.

In order to do this, a new NamedCompositeElement base class class was
introduced which stores a composite name and a list of contained
elements in a dictionary.