Update page element to Class and add serial option
The page element dictionary has been converted to a Page class.  LIVVkit
produces the correct pages on the output website, but does not yet take
full advantage of the class methods.

In order to do this, the summary dictionaries had to be handled a little
differently and are returned from run_suite component functions instead
of written to a shared memory dictionary (better practices anyways) and
the multiprocessing method has switched from starting a list of
processes to running them within a pool of workers.

A command line option for `-p/--pool-size` was added so the user can
select the size of the processing pool, where a pool size of zero causes
a serial run. This changes the `-p` option, removing the under utilized
`-p/--publish` option that added publication quality figures because
that methodology wasn't automatic or easy for users to implement.

Similar functionality would be better used through a figure option which
allows you to specify figure properties automatically (though this will
be hard to implement in LEX as users are free to use matplotlib or not).